fix(nse): resolve bugs and clean up slapper-nse module
- Replace deprecated is_path_allowed() with get_allowed_path() in io.rs
- Remove duplicate CveClient trait from cve/mod.rs, re-export from traits.rs
- Remove duplicate get_builtin_script() from nse_tool.rs
- Delete dead code files peg_parser.rs and pest_bridge.rs (1146 lines)
- Auto-fix clippy warnings across 50+ library files
- Fix webhook.rs serialization refactor

+++ crates/slapper-nse/src/libraries/kafka.rs
@@ -53,7 +53,7 @@ impl KafkaConnection {
     ) -> Result<Vec<u8>, String> {
         let mut buffer = Vec::new();
 
-        let message_len = 4 + 2 + 2 + 4 + request.len() + 2 + CLIENT_ID.len() as usize;
+        let message_len = 4 + 2 + 2 + 4 + request.len() + 2 + CLIENT_ID.len();
 
         buffer.extend_from_slice(&(message_len as i32).to_be_bytes());
         buffer.extend_from_slice(&api_key.to_be_bytes());
